refactored sighting class organisation
