--- src/main/java/com/example/trainspottingjava/sighting/SightingService.java
@@ -1,4 +1,0 @@
-package com.example.trainspottingjava.sighting;
-
-public class SightingService {
-}
