+++ src/main/java/com/example/trainspottingjava/sighting/repository/SightingRepository.java
@@ -1,0 +1,4 @@
+package com.example.trainspottingjava.sighting.repository;
+
+public class SightingRepository {
+}
