+++ src/main/java/com/example/trainspottingjava/sighting/service/SightingService.java
@@ -1,0 +1,4 @@
+package com.example.trainspottingjava.sighting.service;
+
+public class SightingService {
+}
